Reuse `unix::cmath`
